Ensure that newlines are split in merged contents
Also refactors Frame drawing to specify the location of any horizontal
offsets directly in the buffer write.

But there is currently no way to space out the lines without processing
them individually and adding string padding. This seems weird, but may
be better than using a bunch of different writes to screenbuffer?

May move to individual screenbuffer writes in the future if there is a
layering or transparency concern.