Fix columnar arrangement - use hoffsets
Now it gets the list of nested contents and their size ahead of time and
preprocesses it into a single number to use to pad out the gaps between.

This only needs to happen with the nested content is not a Frame with a
set size other than `:none`. The rest of the time, this code doesn't
really do anything.

At least it shouldn't!