Change expected behavior based on last commit
Previously, the code would grab the last line and then append the bottom
Frame to it. That was defensivble bahvior for this edge case.

Now, it adds a new line to the end and adds it there. I think this is
also defensible.

I consider this an edge case that I have spent too much time on already,
  but ideally I think that it should truncate the top of the nested
  Frame and its bottom line should be on the bottom line of the parent
  Frame, instead of the top line of the nested Frame being on the bottom
  line of the parent Frame.